requirements.txt
langchain-anthropic>=0.1.0
langgraph>=0.0.30
pydantic>=2.0
python-dotenv>=1.0.0
requests>=2.31.0
streamlit>=1.30.0
